Use separate view for general metrics
